extract content service
