+++ app/Contracts/ContentServiceInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Contracts;
+
+use Filament\Forms\Components\Builder;
+
+interface ContentServiceInterface
+{
+    public function getFilamentContentBuilder(string $name): Builder;
+}
