+++ app/Filament/Actions/Base/CustomFilamentAction.php
@@ -3,13 +3,12 @@
 namespace App\Filament\Actions\Base;
 
 use Closure;
-use Filament\Actions\StaticAction;
 use Filament\Tables\Actions\Action;
 use Illuminate\Database\Eloquent\Model;
 
 abstract class CustomFilamentAction implements CustomFilamentActionInterface
 {
-    final public static function make(): StaticAction
+    final public static function make(): Action
     {
         return Action::make(static::getActionName())
             ->visible(static::canRun())
