+++ app/Filament/Actions/Base/CustomFilamentBulkAction.php
@@ -3,13 +3,12 @@
 namespace App\Filament\Actions\Base;
 
 use Closure;
-use Filament\Actions\StaticAction;
 use Filament\Tables\Actions\BulkAction;
 use Illuminate\Database\Eloquent\Collection;
 
 abstract class CustomFilamentBulkAction implements CustomFilamentActionInterface
 {
-    final public static function make(): StaticAction
+    final public static function make(): BulkAction
     {
         return BulkAction::make(static::getActionName())
             ->visible(static::canRun())
