+++ app/Providers/AppServiceProvider.php
@@ -3,9 +3,11 @@
 namespace App\Providers;
 
 use App\Contracts\BreadcrumbsServiceInterface;
+use App\Contracts\ContentServiceInterface;
 use App\Contracts\InterlinkingServiceInterface;
 use App\Contracts\RedirectServiceInterface;
 use App\Services\BreadcrumbsService;
+use App\Services\ContentService;
 use App\Services\InterlinkingService;
 use App\Services\RedirectService;
 use Illuminate\Database\Eloquent\Model;
@@ -15,6 +17,7 @@ class AppServiceProvider extends ServiceProvider
 {
     public array $bindings = [
         BreadcrumbsServiceInterface::class => BreadcrumbsService::class,
+        ContentServiceInterface::class => ContentService::class,
         RedirectServiceInterface::class => RedirectService::class,
         InterlinkingServiceInterface::class => InterlinkingService::class,
     ];
